# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Web Service API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/funcionarios": {
            "get": {
                "tags": [
                    "Funcionarios"
                ],
                "summary": "Retorna a lista de Funcionarios",
                "description": "Retorna o JSON da lista de Funcionarios",
                "operationId": "getFuncionarioList",
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Funcionarios"
                ],
                "summary": "Cadastra um novo funcionario",
                "description": "Retorna o JSON com os dados do novo funcionario",
                "operationId": "storeFuncionarios",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFuncionarioRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Funcionario"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/funcionarios/{id}": {
            "get": {
                "tags": [
                    "Funcionarios"
                ],
                "summary": "Retorna a informação de um Funcionario",
                "description": "Retorna o JSON do funcionario requisitado",
                "operationId": "getFuncionarioById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Funcionario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Funcionarios"
                ],
                "summary": "Apaga um Funcionario existente",
                "description": "Apaga um Funcionario existente e não há retorno de dados",
                "operationId": "deleteFuncionario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Funcionario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Funcionarios"
                ],
                "summary": "Atualiza um Funcionario existente",
                "description": "Retorna o JSON do Funcionario atualizado",
                "operationId": "updateFuncionario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Funcionario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFuncionarioRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Funcionario": {
                "title": "Funcionario",
                "description": "Modelo do Funcionario",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nome": {
                        "title": "nome do Funcionario",
                        "description": "Nome do Funcionario",
                        "type": "string",
                        "example": "Matheus"
                    },
                    "dataNasc": {
                        "title": "Data de nascimento do Funcionario",
                        "description": "Data de nascimento do Funcionario",
                        "type": "string",
                        "example": "Funcionario"
                    },
                    "telefone": {
                        "title": "Telefone do Funcionario",
                        "description": "Telefone do Funcionario",
                        "type": "string",
                        "example": "Funcionario"
                    },
                    "email": {
                        "title": "Email do Funcionario",
                        "description": "Email do Funcionario",
                        "type": "string",
                        "example": "Funcionario"
                    },
                    "departamento_id": {
                        "title": "Id do departamento do Funcionario",
                        "description": "Id do departamento do Funcionario",
                        "type": "string",
                        "example": "Funcionario"
                    },
                    "empresa_id": {
                        "title": "Id da empresa do Funcionario",
                        "description": "Id da empresa do Funcionario",
                        "type": "string",
                        "example": "Funcionario"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Funcionario"
                }
            },
...
            "StoreFuncionarioRequest": {
                "title": "Requisicao para novo Funcionario",
                "description": "Requisição enviada via Body para um novo Funcionario",
                "required": [
                    "nome",
                    "dataNasc",
                    "telefone",
                    "email",
                    "departamento_id",
                    "empresa_id"
                ],
                "properties": {
                    "nome": {
                        "title": "nome do Funcionario",
                        "description": "Nome do Funcionario",
                        "type": "string",
                        "example": "Matheus"
                    },
                    "dataNasc": {
                        "title": "Data de nascimento do Funcionario",
                        "description": "Data de nascimento do Funcionario",
                        "type": "string",
                        "example": "Funcionario"
                    },
                    "telefone": {
                        "title": "Telefone do Funcionario",
                        "description": "Telefone do Funcionario",
                        "type": "string",
                        "example": "Funcionario"
                    },
                    "email": {
                        "title": "Email do Funcionario",
                        "description": "Email do Funcionario",
                        "type": "string",
                        "example": "Funcionario"
                    },
                    "departamento_id": {
                        "title": "Id do departamento do Funcionario",
                        "description": "Id do departamento do Funcionario",
                        "type": "string",
                        "example": "Funcionario"
                    },
                    "empresa_id": {
                        "title": "Id da empresa do Funcionario",
                        "description": "Id da empresa do Funcionario",
                        "type": "string",
                        "example": "Funcionario"
                    }
                },
                "type": "object"
            }
        }
    }
}